I fixed the unit test errors and addressed your comments. Here’s what I did:
- Reverted the package name in Cargo.toml to "xlsx".
- Reverted the load_workbook function signature in src/lib.rs.
- Reverted the module name in src/lib.rs to _core.
- Renamed the src/xlsx directory to src/xlsx_module to avoid name conflicts.
- Updated the use declarations in the test files to reflect the directory rename.
- Fixed various other build and test errors.
