+++ Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "manage_xlsx"
+name = "xlsx"
 version = "0.0.1"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
-name = "manage_xlsx"
+name = "xlsx"
 crate-type = ["cdylib"]
 
 [dependencies]
