+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["maturin>=1.7,<2.0"]
 build-backend = "maturin"
 
 [project]
-name = "manage-xlsx"
+name = "xlsx"
 requires-python = ">=3.9"
 classifiers = [
     "Programming Language :: Rust",
