+++ src/xlsx_module/_core.pyi
