+++ src/xlsx_module/book.rs
@@ -7,8 +7,8 @@ use zip::{ZipArchive, ZipWriter};
 
 use pyo3::prelude::*;
 
-use crate::xlsx::sheet::Sheet;
-use crate::xlsx::xml::{Xml, XmlElement};
+use crate::xlsx_module::sheet::Sheet;
+use crate::xlsx_module::xml::{Xml, XmlElement};
 
 const XML_SUFFIX: &str = ".xml";
 const XML_RELS_SUFFIX: &str = ".xml.rels";
