+++ src/xlsx_module/cell.rs
@@ -1,5 +1,5 @@
-use crate::xlsx::style::{Font, PatternFill};
-use crate::xlsx::xml::{Xml, XmlElement};
+use crate::xlsx_module::style::{Font, PatternFill};
+use crate::xlsx_module::xml::{Xml, XmlElement};
 use chrono::{NaiveDateTime};
 use pyo3::prelude::*;
 use std::sync::{Arc, Mutex};
